finished error handling
