+++ client/src/styles/_main.scss
@@ -11,19 +11,23 @@ body {
     margin: 0 auto;
 }
 
-#error_message {
+.error_message {
     background-color: $errorMessage-color;
     padding: 5px;
-    width: 200px;
-    height: 150px;
+    max-width: 1000px;
+    width: auto;
+    height: auto;
     margin: 10px;
-    line-height: 150px;
     color: $errorMessageText-color;
-    font-weight: bold;
-    font-size: 3em;
-    text-align: center;
+    overflow: scroll;
 }
 
 #footer {
     margin-top: 50px;
+}
+
+@media screen and (max-width: 500px) {
+    #appContainer {
+        width: 500px;
+    }
 }
