+++ client/src/views/_error-view.html
@@ -1,5 +1,7 @@
 <div class="container-flex">
-    <div id="error_message">
-        Error
+    <div class="error_message">
+        <pre>
+{{msg ? msg : 'ERROR 404'}}
+        </pre>
     </div>
 </div>
