+++ package.json
@@ -43,6 +43,7 @@
   "dependencies": {
     "body-parser": "1.17.1",
     "express": "4.15.2",
+    "q": "1.5.0",
     "underscore": "1.8.3"
   }
 }
